@@ -1,10 +1,6 @@
 apply plugin: 'com.android.library'
-
 apply plugin: 'kotlin-android'
-
 apply plugin: 'kotlin-kapt'
-
-//apply plugin: 'com.novoda.bintray-release'
 
 buildscript {
     apply from: '../versions.gradle'
@@ -15,22 +11,14 @@ buildscript {
     dependencies {
         classpath "com.android.tools.build:gradle:$gradle_tools_version"
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
-       // classpath "com.novoda:bintray-release:$novoda_bintray_version"
     }
 }
 
 android {
-    compileSdkVersion library_compile_version
-    buildToolsVersion library_build_tools_version
-
     defaultConfig {
         minSdkVersion library_min_version
         targetSdkVersion library_target_version
-        versionCode library_version_code
-        versionName library_version
-
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
-
+        compileSdk library_compile_version
     }
 
     buildTypes {
@@ -44,32 +32,35 @@ android {
 
         }
     }
-
-    lintOptions {
-        abortOnError false
+    buildFeatures {
+        dataBinding = true
+        buildConfig = true
     }
 
+
     compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_7
-        targetCompatibility JavaVersion.VERSION_1_7
+        sourceCompatibility JavaVersion.VERSION_17
+        targetCompatibility JavaVersion.VERSION_17
     }
 
     sourceSets {
         main.java.srcDirs += 'src/main/kotlin'
         androidTest.java.srcDirs += 'src/androidTest/kotlin'
     }
+    namespace 'com.tonyodev.fetch2'
+    lint {
+        abortOnError false
+    }
 }
 
 dependencies {
 
     api "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
-    kapt "android.arch.persistence.room:compiler:$room_version"
-    api "android.arch.persistence.room:runtime:$room_version"
+    implementation "androidx.room:room-runtime:$room_version"
+    implementation "androidx.room:room-ktx:$room_version"
+    implementation 'androidx.core:core-ktx:1.13.1'
+    kapt "androidx.room:room-compiler:$room_version"
     api project(':fetch2core')
-
-    testImplementation "junit:junit:$junit_version"
-    androidTestImplementation "com.android.support.test:runner:$test_runner_version"
-    androidTestImplementation "com.android.support.test.espresso:espresso-core:$espresso_version"
 }
 
 //publish {
