@@ -1,8 +1,6 @@
 apply plugin: 'com.android.library'
-
 apply plugin: 'kotlin-android'
-
-//apply plugin: 'com.novoda.bintray-release'
+apply plugin: 'kotlin-kapt'
 
 buildscript {
     apply from: '../versions.gradle'
@@ -13,21 +11,14 @@ buildscript {
     dependencies {
         classpath "com.android.tools.build:gradle:$gradle_tools_version"
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
-//        classpath "com.novoda:bintray-release:$novoda_bintray_version"
     }
 }
 
 android {
-    compileSdkVersion library_compile_version
-
     defaultConfig {
         minSdkVersion library_min_version
         targetSdkVersion library_target_version
-        versionCode library_version_code
-        versionName library_version
-
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
-
+        compileSdk library_compile_version
     }
 
     buildTypes {
@@ -36,19 +27,24 @@ android {
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
     }
-
-    lintOptions {
-        abortOnError false
+    buildFeatures {
+        dataBinding = true
+        buildConfig = true
     }
 
+
     compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_7
-        targetCompatibility JavaVersion.VERSION_1_7
+        sourceCompatibility JavaVersion.VERSION_17
+        targetCompatibility JavaVersion.VERSION_17
     }
 
     sourceSets {
         main.java.srcDirs += 'src/main/kotlin'
         androidTest.java.srcDirs += 'src/androidTest/kotlin'
+    }
+    namespace 'com.tonyodev.fetch2rx'
+    lint {
+        abortOnError false
     }
 }
 
@@ -59,10 +55,6 @@ dependencies {
     api "io.reactivex.rxjava2:rxandroid:$rxAndroid2_version"
     api "io.reactivex.rxjava2:rxjava:$rxJava2_version"
     api project(':fetch2')
-
-    testImplementation "junit:junit:$junit_version"
-    androidTestImplementation "com.android.support.test:runner:$test_runner_version"
-    androidTestImplementation "com.android.support.test.espresso:espresso-core:$espresso_version"
 }
 
 //publish {
