@@ -1,22 +1,22 @@
 ext {
-    kotlin_version = '1.3.72'
+    kotlin_version = '1.9.24'
     okhttp_version = '3.12.12'
     okhttp_url_version = '3.12.12'
-    android_support_version = '28.0.0'
     constraint_layout_version = "1.1.3"
-    room_version = '1.1.1'
-    junit_version = '4.12'
+    room_version = '2.5.1'
     espresso_version = '3.0.1'
     test_runner_version = '1.0.1'
-    library_min_version = '14'
-    library_compile_version = 30
-    library_target_version = 30
-    library_build_tools_version = "30.0.3"
-    gradle_tools_version = '3.4.1'
+    library_min_version = '28'
+    library_compile_version = 34
+    library_target_version = 34
+    gradle_tools_version = '8.5.2'
     rxJava2_version = "2.2.20"
     rxAndroid2_version = "2.1.1"
     timber_version = "4.7.1"
-    novoda_bintray_version = "0.9.2"
-    library_version = "3.0.12"
-    library_version_code = 77
+    library_version = "3.2.0"
+    library_version_code = 100
+
+    appcompat_version = '1.7.0'
+    materialLibVersion = '1.12.0'
+    constraintLayoutVersion = '2.1.4'
 }
