@@ -8,7 +8,8 @@ import android.content.Intent
 import android.content.IntentFilter
 import android.net.Uri
 import android.os.Build
-import android.support.v4.app.NotificationCompat
+import androidx.core.app.NotificationCompat
+import androidx.core.content.ContextCompat
 
 import com.tonyodev.fetch2.DownloadNotification.ActionType.*
 import com.tonyodev.fetch2.util.DEFAULT_NOTIFICATION_TIMEOUT_AFTER
@@ -53,7 +54,7 @@ abstract class DefaultFetchNotificationManager(context: Context) : FetchNotifica
     }
 
     override fun registerBroadcastReceiver() {
-        context.registerReceiver(broadcastReceiver, IntentFilter(notificationManagerAction))
+        ContextCompat.registerReceiver(context, broadcastReceiver, IntentFilter(notificationManagerAction), ContextCompat.RECEIVER_EXPORTED)
     }
 
     override fun unregisterBroadcastReceiver() {
