@@ -3,7 +3,7 @@ package com.tonyodev.fetchapp;
 import android.content.Context;
 import android.net.Uri;
 import android.os.Environment;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 import com.tonyodev.fetch2.Priority;
 import com.tonyodev.fetch2.Request;
