@@ -1,10 +1,10 @@
 package com.tonyodev.fetch2.database
 
-import android.arch.persistence.db.SupportSQLiteDatabase
-import android.arch.persistence.room.Room
+import androidx.room.Room
 import android.content.Context
 import android.database.Cursor
 import android.database.sqlite.SQLiteException
+import androidx.sqlite.db.SupportSQLiteDatabase
 import com.tonyodev.fetch2.PrioritySort
 import com.tonyodev.fetch2.Status
 import com.tonyodev.fetch2.database.migration.Migration
