@@ -1,6 +1,6 @@
 package com.tonyodev.fetch2.database.migration
 
-import android.arch.persistence.db.SupportSQLiteDatabase
+import androidx.sqlite.db.SupportSQLiteDatabase
 import com.tonyodev.fetch2.database.DownloadDatabase
 import com.tonyodev.fetch2.util.DEFAULT_AUTO_RETRY_ATTEMPTS
 
