@@ -1,9 +1,9 @@
 package com.tonyodev.fetch2fileserver.database
 
-import android.arch.persistence.room.ColumnInfo
-import android.arch.persistence.room.Entity
-import android.arch.persistence.room.Index
-import android.arch.persistence.room.PrimaryKey
+import androidx.room.ColumnInfo
+import androidx.room.Entity
+import androidx.room.Index
+import androidx.room.PrimaryKey
 
 @Entity(tableName = FileResourceInfoDatabase.TABLE_NAME,
         indices = [(Index(value = [FileResourceInfoDatabase.COLUMN_NAME], unique = true)),
