@@ -1,7 +1,7 @@
 package com.tonyodev.fetch2fileserver.database
 
-import android.arch.persistence.room.Database
-import android.arch.persistence.room.RoomDatabase
+import androidx.room.Database
+import androidx.room.RoomDatabase
 import com.tonyodev.fetch2fileserver.database.FileResourceInfoDatabase.Companion.DATABASE_VERSION
 
 @Database(entities = [FileResourceInfo::class], version = DATABASE_VERSION, exportSchema = false)
